update server kdw008

+++ nts.uk/uk.at/at.ctx/function/nts.uk.ctx.at.function.app/src/main/java/nts/uk/ctx/at/function/app/command/modifyanyperiod/ModifyAnyPeriodDetailDto.java
@@ -17,5 +17,5 @@ public class ModifyAnyPeriodDetailDto {
     private int order;
 
 
-    private BigDecimal columnWidth;
+    private Integer columnWidth;
 }
